Area/Region Rework
I've reworked the areas/regions/superregions to mirror the setup the
state system was designed for, i.e. areas of 3-10 provinces. This
reduces the size of states considerably.

I've also altered all instances where specific areas/regions/etc are
referred to, to fit the new setup.

The groups for discovery are now provincegroups