Horde Conquest
# Removed the province shifting and depopulation events (sorry Moridin).
These were buggy and were too broad.
# Removed the orc raiding event.
# Added a Conquest event for Scourge/Warband governments, that fires
when they siege down a enemy province. This let's them choose between
Raising the dead (scourge only), converting culture and religion in
exchange for unrest, taking and coring the province, in exchange for
very high unrest, or doing nothing.
This is to replicate what Moridin wrote, but in a way that is
predictable and learnable by the player.
# Added a decision allowing the player to set their stance on the
Conquest event, letting them set the event to auto-resolve to a chosen
decision.
# Removed the forced wait before you can declare war on game start
# Fixed a bug with provincegroup not being province_group in the
Alliance/Horde decisions
# Fixed a bug with area_durotar not being region_durotar in wow_story
# Added slower technology modifiers for a few technology groups